requirements.txt
gradio>=6.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pillow>=9.0.0
transformers>=4.40.0
torch>=2.0.0